pyproject.toml
[project]
name = "robi-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "deepagents>=0.4.3",
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
    "websockets<15",
    "wsproto>=1.3.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "streamlit>=1.54.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
